[move-stdlib] rename `std::string` and `std::ascii` `length` function to `size`
The current implementation of `std::string` and `std::ascii` `length` returns the number of bytes of the string and not the number of characters that makes up the string.

This can be misleading and confusing to developers as most languages implementation of the string's `length` function returns the number of characters and not bytes.

This commit does:
- rename the `length` function to `size`
- update the test files to use `size` instead of `length`
- generate documentation to reflect the changes